requirements.txt
beautifulsoup4
python-Levenshtein